requirements.txt
flask
openai
azure-cognitiveservices-speech
